+++ src/generators/changelogPage.js
@@ -1,19 +1,25 @@
 const FlexSearch = require('flexsearch')
-
 const fs = require('fs')
 const env = require('jsdoc/env')
 const path = require('path')
+const shell = require('shelljs')
 const md = require('markdown-it')()
 // autocreate anchor from headers #...####
-md.use(require('markdown-it-anchor'))
+md.use(require('markdown-it-anchor'), {
+  // for preventing change @ and / to %xx
+  // assumed that package name already slugified
+  slugify: s => s
+})
   // for :!: emoji and not only
   .use(require('markdown-it-emoji'))
   // for smaller date and version (use ~text~ in .md)
   .use(require('markdown-it-sub'))
 const renderFile = require('../vueRender')
-const { idGeneratorFabric } = require('../utils')
-const { createItemFileName } = require('../utils')
-const { getChangelogPaths, getParsedChangelogs, getParseErrors, groupingChanges, filterLogByDate, renderToMD } = require('ub-changelog-parser')
+const { idGeneratorFabric, createItemFileName } = require('../utils')
+const {
+  getChangelogPaths, getParsedChangelogs, getParseErrors,
+  groupingChanges, filterLogByDate, sortPackagesByNames, renderToMD
+} = require('ub-changelog-parser')
 
 const outDir = path.normalize(env.opts.destination)
 const isEmptyObj = obj => Object.keys(obj).length === 0
@@ -36,7 +42,7 @@ const createNavigation = (tree, year, month) => Object.keys(tree)
     })
   )
 
-/// search
+// search
 const index = new FlexSearch({
   doc: {
     id: 'id',
@@ -52,12 +58,12 @@ const addToSearch = (cl, link) => {
   const id = getFTSid()
   index.add({
     id: id,
-    name: cl.pkgName,
-    description: Object.values(cl.versions).map(change => change.map(({ log }) => log.join(' ')).join(' ')).join((' '))
+    name: cl.name,
+    description: cl.changes.map(({ versions }) => versions.map(({ log }) => log.join(' ')).join(' ')).join(' ')
   })
   ftsData[id] = {
-    link: `${link}#${cl.pkgName}`,
-    parent: cl.pkgName
+    link: `${link}#${cl.name}`,
+    parent: cl.name
   }
 }
 
@@ -82,8 +88,13 @@ const changelog = () => {
     for (let month = 0; month <= endMonth; month++) {
       const fromDate = new Date(year, month, 1)
       const toDate = new Date(year, month + 1, 0)
-      const monthCl = parsedChangelogs.map(cl => groupingChanges(filterLogByDate(cl, fromDate, toDate)))
-        .filter(({ versions }) => !isEmptyObj(versions))
+      const monthCl = parsedChangelogs
+        .map(cl => filterLogByDate(cl, fromDate, toDate))
+        .map(cl => groupingChanges(cl))
+        .filter(({ changes }) => changes.length > 0)
+      if (config.order) {
+        monthCl.sort((a, b) => sortPackagesByNames(a.name, b.name, config.order))
+      }
       // prevent empty month pages
       if (monthCl.length !== 0) {
         monthTree[month + 1] = monthCl
@@ -98,34 +109,14 @@ const changelog = () => {
   if (!fs.existsSync(path.resolve(outDir, '../changelog'))) {
     fs.mkdirSync(path.resolve(outDir, '../changelog'))
   }
-  const renderNavigation = navigation =>
-    '# Monthly changes\n' + navigation
-      .map(({ name, submenu }) => `#### ${name}\n` + submenu
-        .map(({ link, name: mName }) => `* [${mName}](${link})`)
-        .join('\n'))
-      .join('\n')
-
-  const navigation = createNavigation(changelogDateTree)
-
-  // create index.html
-  const indexHtml = md.render(renderNavigation(navigation))
-  renderFile(
-    {
-      navigation,
-      html: indexHtml,
-      tableOfContent: []
-    },
-    path.resolve(__dirname, '../../tmpl/vue/gettingStarted.vue'),
-    path.resolve(__dirname, '../../tmpl/html/pageTemplate.html'),
-    path.resolve(outDir, '../changelog', 'index.html')
-  )
 
   Object.entries(changelogDateTree).forEach(([year, months]) => {
     Object.entries(months).forEach(([month, cls]) => {
-      const menu = cls.map(cl => cl.pkgName).map(pkgName => ({ name: pkgName, link: `#${pkgName}` }))
+      const menu = cls.map(({ name }) => ({ name, link: `#${name}` }))
       const tableOfContent = [{ name: 'Packages', props: menu }]
+      // if (year === '2020') {debugger} else return
       const html = md.render(renderToMD(cls))
-      const fileName = createItemFileName('cl', year + '-' + month)
+      const fileName = createItemFileName('cl', `${year}-${month}`)
       renderFile(
         {
           navigation: createNavigation(changelogDateTree, year, month),
@@ -140,6 +131,14 @@ const changelog = () => {
       cls.forEach(cl => addToSearch(cl, fileName))
     })
   })
+
+  // set index.html as **last present** month and year changelog to simplify ci
+  // we can't get current year or month because it may not contain changes
+  const lastPresentYear = Math.max(...Object.keys(changelogDateTree).map(Number))
+  const lastPresentMonth = Math.max(...Object.keys(changelogDateTree[String(lastPresentYear)]).map(Number))
+  shell.cp('-f',
+    path.resolve(outDir, '../changelog', createItemFileName('cl', `${lastPresentYear}-${lastPresentMonth}`)),
+    path.resolve(outDir, '../changelog', 'index.html'))
 
   fs.writeFileSync(path.resolve(outDir, '../changelog', 'ftsIndex.json'), JSON.stringify(index.export()))
   fs.writeFileSync(path.resolve(outDir, '../changelog', 'ftsData.json'), JSON.stringify(ftsData))