+++ tmpl/vue/elements/anchor.vue
@@ -1,3 +1,2 @@
-<!--<span class="anchor">#</span>-->
 <a :data-id="id" tabindex="0" class="anchor" role="button" data-toggle="popover" data-trigger="focus" data-placement="top" data-animation="false"
    data-content="Copied!">#</a>