+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+## [2.3.2]
+### Added
+ - ordering for packages in changelog 
+### Changed
+ - index page in changelog now copy of current month changelog page 
+
 ## [2.3.1]
 ### Added
  - index page for changelog generator